Add config option to ignore TLS cert verification
This allows users to disable the TLS cert checks performed
by requests when using the provisioner API (at their own
risk).

The main reason for supporting this is to support trial
installs where a self-signed cert might be used as a placeholder
while the user decides if OpenLI is for them.

They can (and should!) switch over to a cert signed by a
proper CA afterwards and then they won't need this option any
more.