Add encryption config options to web UI
Do not "trim" fields with empty strings as values before creating
an object to send to the provisioner -- an empty string may be a
valid value (e.g. as an encryption key).

Added custom validation for encryption keys -- they're only
required if the encryption method is set to anything other than
"none", so we have to check both data fields to validate whether
the key (or lack thereof) is allowed.