Fix empty "realm" bug when adding VoIP intercepts
Intended behaviour: leaving the realm field on the intercept
creation form blank would lead to a submitted target object
that contained the username alone.

Actual behaviour: the submitted target object would contain a field
`"realm": ""`, which OpenLI would not treat as a wildcard (although
perhaps it should). Since no sane SIP URL will have an empty
string in place of a realm, the submitted target would never be
matched in the SIP traffic seen by OpenLI.

This commit updates the form submission code to remove any
fields with empty string values IF the field is defined in api.yml
as having a "true" value for the newly-added "dropblank"
property. So far, only "realm" within a SIP target definition has
this property defined.