Ensure that the correct agency is selected when editing an intercept
The agency selector would always default to the "first" agency
in the list, even if the intercept had been configured to use
a different agency.

This meant that the agency would get reset if the user edited any
other intercept properties. Also, it is incredibly confusing for
users to browse intercepts via the "edit" facility and see that
their agency appears to be set incorrectly (when in fact, it is
not but we are telling them otherwise).